Move expression.resolveLocality to ExpressionLocalityResolver
